Renaming test objects
